amd-pstate: Validate writes work during init
Due a firmware bug on some OEM systems the amd-pstate kernel driver
starts up in active mode but EPP writes don't work.

Admittedly this is a kernel bug that it still advertises support and
it could be handled better by the kernel, but PPD can also validate
writes with a no-op for each CPU at bootup.

Do this to ensure that the CPU driver is only setup if we know it works.