ci: Use ubuntu images for CI jobs
It's the distro I'm using, so it makes things easier to maintain.

It tend to be also a bit more stable, so it helps unless we need
something really bleeding edge (but we can still switch to devel if
required).

Finally fedora does not use ppd anymore, so there's no much point to
use it as our development target